fix: resolve release build crash + type mismatches in FRB bindings
- Fix AuthCodeResult mirror types: code_length u32, resend_after_seconds u64
- Make check_connection async to avoid Tokio reactor panic in release builds
- Remove unused imports from onboarding, settings, upload screens
- Update .gitignore: exclude jniLibs and screenshots from tracking
- Regenerate FRB bindings with corrected types
- Verified: APK launches clean on emulator (x86_64), no runtime errors

+++ app_flutter/rust/src/api/mirror.rs
@@ -16,8 +16,8 @@ pub use telegram_photos_core::models::{
 #[frb(mirror(AuthCodeResult))]
 pub struct _AuthCodeResult {
     pub status: String,
-    pub code_length: Option<i32>,
-    pub resend_after_seconds: Option<i32>,
+    pub code_length: Option<u32>,
+    pub resend_after_seconds: Option<u64>,
     pub delivery: Option<String>,
 }
